Add files via upload
This is a guided project I have completed during my time working through the Data Analysis in Python track from Dataquest. Throughout the guided project, what I did was look for the information that the project required, but used my own methodology based on what was learnt to be able to compute the data to get the desired results. the CSV that was used in the data analysis is included in this folder